NewAPP/requirements.txt
flask
flask-cors
transformers
python-dotenv
mysql-connector-python
PyJWT